fix(docs): isolate API generate from Vue JSX typecheck
vitepress-api-references pulls VitePress/Vue types into the unit
typecheck program and breaks vue-i18n-core component JSX tests.
Keep generate helpers in a lib, run generate via CLI, and typecheck
the VitePress-facing files in a separate tsconfig.
