+++ scripts/api-docs.config.ts
