+++ scripts/generate-docs.test.ts
@@ -1,14 +1,12 @@
-import { access, readFile } from 'node:fs/promises'
+import { execFileSync } from 'node:child_process'
+import { access, readFile, readdir } from 'node:fs/promises'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import {
-  collectGeneratedSymbolNames,
-  generateApiDocs,
   isAllowedDiagnostic,
   normalizeGeneratedFiles,
   removeModulesSection
-} from './generate-docs'
-import { syncApiDocsLocales } from './generate-api-locales'
+} from './generate-docs-lib'
 import baseline from './api-docs-symbol-baseline.json' with { type: 'json' }
 
 const rootDir = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..')
@@ -20,6 +18,23 @@ const requiredPages = [
   'Path',
   'VERSION'
 ] as const
+
+async function walkMarkdown(dir: string, acc: string[] = []): Promise<string[]> {
+  const entries = await readdir(dir, { withFileTypes: true })
+  for (const entry of entries) {
+    const fullPath = path.join(dir, entry.name)
+    if (entry.isDirectory()) {
+      if (entry.name !== 'v11') {
+        await walkMarkdown(fullPath, acc)
+      }
+      continue
+    }
+    if (entry.name.endsWith('.md') && entry.name !== 'index.md') {
+      acc.push(fullPath)
+    }
+  }
+  return acc
+}
 
 describe('generateApiDocs', () => {
   test('skips TypeDoc-style /modules/ pages', () => {
@@ -56,25 +71,25 @@ describe('generateApiDocs', () => {
   })
 
   test('writes the TypeDoc public symbol set without touching v11 pages', async () => {
-    const { files, diagnostics, outDir } = await generateApiDocs()
-    const names = collectGeneratedSymbolNames(files)
+    execFileSync('pnpm', ['docs:api'], { cwd: rootDir, stdio: 'inherit' })
+    const firstPages = await walkMarkdown(path.join(rootDir, 'docs/api'))
+    const names = [...new Set(firstPages.map(filePath => path.basename(filePath, '.md')))].sort()
 
     expect(names).toEqual([...baseline].sort())
     expect(names).not.toContain('ComponentCustomOptions')
-    expect(diagnostics.filter((message: string) => !isAllowedDiagnostic(message))).toEqual([])
 
     for (const name of requiredPages) {
-      const filePath = Object.keys(files).find(entry => entry.endsWith(`/${name}.md`))
+      const filePath = firstPages.find(entry => entry.endsWith(`${path.sep}${name}.md`))
       expect(filePath, `${name} page`).toBeTruthy()
-      expect(files[filePath!].trim().length).toBeGreaterThan(0)
+      expect((await readFile(filePath!, 'utf8')).trim().length).toBeGreaterThan(0)
     }
 
-    await access(path.join(outDir, 'v11/general.md'))
-    await generateApiDocs()
-    const second = await readFile(path.join(outDir, 'v11/general.md'), 'utf8')
+    await access(path.join(rootDir, 'docs/api/v11/general.md'))
+    execFileSync('pnpm', ['docs:api'], { cwd: rootDir, stdio: 'inherit' })
+    const second = await readFile(path.join(rootDir, 'docs/api/v11/general.md'), 'utf8')
     expect(second.length).toBeGreaterThan(0)
 
-    await syncApiDocsLocales()
+    execFileSync('pnpm', ['docs:api:locales'], { cwd: rootDir, stdio: 'inherit' })
     const jpSidebar = await readFile(path.join(rootDir, 'docs/jp/api/api-sidebar.json'), 'utf8')
     const zhSidebar = await readFile(path.join(rootDir, 'docs/zh/api/api-sidebar.json'), 'utf8')
     expect(jpSidebar).toContain('/jp/api/')
